% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Table_Description.R
\docType{data}
\name{BRACKETS}
\alias{BRACKETS}
\title{Brackets for the fhir table descriptions used for the indices of multi value entries.}
\format{
An object of class \code{character} of length 2.
}
\usage{
BRACKETS
}
\description{
Brackets for the fhir table descriptions used for the indices of multi value entries.
}
\details{
If a xpath expression returns more than 1 value (if it is a list of multiple values (e.g.
list of codes and corresponding code systems), then all
}
\keyword{datasets}
