% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{convertFilterPatterns}
\alias{convertFilterPatterns}
\title{Converts a filter pattern list from the toml file into an internal representation
as a list of lists. Every subcondition in a sublist must be fulfilled to fulfill
the whole condition represented by the sublist (AND connected). Lines of the table
can be accepted by the filter if at least one of the main conditions (which consists
of these subconditions) in the main list is fulfilled (OR connected).}
\usage{
convertFilterPatterns(
  filter_patterns_global_variable_name_prefix = "ENCOUNTER_FILTER_PATTERN"
)
}
\arguments{
\item{filter_patterns_global_variable_name_prefix}{name of the variable in the glogbal environment which
contains the filter patterns from the toml file}
}
\value{
the filter patterns which are converted to a list of lists
}
\description{
Converts a filter pattern list from the toml file into an internal representation
as a list of lists. Every subcondition in a sublist must be fulfilled to fulfill
the whole condition represented by the sublist (AND connected). Lines of the table
can be accepted by the filter if at least one of the main conditions (which consists
of these subconditions) in the main list is fulfilled (OR connected).
}
