% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Table_Description.R
\name{getTableDescription}
\alias{getTableDescription}
\title{Get a list of fhircrackr::fhir_table_description() objects based on a table description.}
\usage{
getTableDescription(table_description_table = NA)
}
\arguments{
\item{table_description_table}{A data table containing information about
resources, column names, and FHIR expressions. If not provided, it
defaults to the result of calling \code{getTableDescriptionTable()}.}
}
\value{
A list of \code{fhircrackr::fhir_table_description()} objects, where each object
represents a resource group, and its elements correspond to the specified columns
and FHIR expressions.
}
\description{
This function takes a table description table, typically obtained from
\code{getTableDescriptionTable()}, and creates a list of
\code{fhircrackr::fhir_table_description()} objects, each corresponding to a resource
group in the table.
}
\examples{
\dontrun{
  # Example usage:
  table_desc <- getTableDescription()
  # Access the fhir_table_description() object for a specific resource group (e.g., "Encounter")
  encounter_description <- table_desc[["Encounter"]]
}

}
\seealso{
\code{\link{getTableDescriptionTable}}, \code{\link[fhircrackr]{fhir_table_description}}
}
\keyword{data}
\keyword{manipulation}
