% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{getTableDescriptionColumnsFromFilterPatterns}
\alias{getTableDescriptionColumnsFromFilterPatterns}
\title{Get FHIR table description based on filter patterns.}
\usage{
getTableDescriptionColumnsFromFilterPatterns(filter_patterns, ...)
}
\arguments{
\item{filter_patterns}{A list of filter patterns, where each pattern is a list of conditions.
Each condition is expected to have named elements representing column names.}

\item{...}{Additional columns to be included in the FHIR table description.}
}
\value{
A FHIR table description object for the 'Encounter' resource with columns based
on the unique names extracted from the filter patterns, including additional columns.
}
\description{
This function takes a list of filter patterns and extracts unique column names
from them to create a FHIR table description for the 'Encounter' resource.
}
