% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Table_Description.R
\name{getTableDescriptionTable}
\alias{getTableDescriptionTable}
\title{Get a table with the Table Descriptions for all fhir resources from Excel file}
\usage{
getTableDescriptionTable()
}
\value{
A data table with the table descriptions.
}
\description{
This function reads the "Table_Description.xlsx" file from the "extdata" directory
in the package and returns the data as a data frame.
}
