% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{initEncounterPeriodToDownload}
\alias{initEncounterPeriodToDownload}
\title{Initialize encounter retrieval period for downloading.}
\usage{
initEncounterPeriodToDownload()
}
\description{
This function initializes the period for downloading encounters based on various conditions.
If debugging period start and end are provided, those values are used. Otherwise, it checks
if encounter data exists. If not, it sets the period in the past based on the initial days;
otherwise, it sets the period based on the usual days. If debugging period start is provided,
it adjusts the start date by subtracting the initial days; otherwise, it uses the current time.
}
