% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{isEncounterTableEmpty}
\alias{isEncounterTableEmpty}
\title{Checks whether this start of the application is the very first start.}
\usage{
isEncounterTableEmpty()
}
\value{
TRUE when this application fills the database for the first time. This is recognized by whether the
Encounter table is still completely empty. If there is already something in this table, FALSE is returned,
as this cannot be the initial start.
}
\description{
Checks whether this start of the application is the very first start.
}
