% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{parseInterpolarPatientIDsPerWardFromFile}
\alias{parseInterpolarPatientIDsPerWardFromFile}
\title{Parse and interpolate patient IDs from a file.}
\usage{
parseInterpolarPatientIDsPerWardFromFile(path_to_PID_list_file)
}
\arguments{
\item{path_to_PID_list_file}{The path to the file containing patient IDs.}
}
\value{
A unique, sorted list of patient IDs.
}
\description{
This function reads patient IDs from a file specified by the provided path.
The patient IDs are then returned as a unique, sorted list.
}
