% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Table_Description.R
\docType{data}
\name{SEP}
\alias{SEP}
\title{Brackets for the fhir table descriptions}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEP
}
\description{
Brackets for the fhir table descriptions
}
\keyword{datasets}
