% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Table_Description.R
\docType{data}
\name{TABLE_DESCRIPTION}
\alias{TABLE_DESCRIPTION}
\title{Create Table-Descriptions and Designs for relevant resources}
\format{
An object of class \code{list} of length 7.
}
\usage{
TABLE_DESCRIPTION
}
\description{
Create Table-Descriptions and Designs for relevant resources
}
\keyword{datasets}
