% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Get_Interpolar_PIDs.R
\name{getInterpolarPatientIDs}
\alias{getInterpolarPatientIDs}
\title{Extracts the Interploar relevant patient IDs from download Encounter resources.}
\usage{
getInterpolarPatientIDs(path_to_PID_list_file)
}
\value{
the Interploar relevant patient IDs
}
\description{
Extracts the Interploar relevant patient IDs from download Encounter resources.
}
