% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Get_Interpolar_PIDs.R
\name{getPIDsPerWard}
\alias{getPIDsPerWard}
\title{Get unique patient IDs per ward based on filter patterns.}
\usage{
getPIDsPerWard(resources, all_wards_filter_patterns)

getPIDsPerWard(resources, all_wards_filter_patterns)
}
\arguments{
\item{resources}{A list of resources to filter.}

\item{all_wards_filter_patterns}{A list of filter patterns, where each element corresponds to a ward.}
}
\value{
A list where each element corresponds to a ward, and the values are unique patient IDs for that ward.

A list where each element corresponds to a ward, and the values are unique patient IDs for that ward.
The list is structured such that each outer list represents a ward, and the inner lists contain
unique patient IDs for that ward.
}
\description{
This function takes a list of resources and a corresponding list of filter patterns for each ward.
It filters the resources for each ward based on the provided filter patterns and extracts unique
patient IDs ('subject/reference'). The result is a list where each element corresponds to a ward,
and the values are unique patient IDs for that ward.

This function takes a list of resources and a corresponding list of filter patterns for each ward.
It filters the resources for each ward based on the provided filter patterns and extracts unique
patient IDs ('subject/reference'). The result is a list where each element corresponds to a ward,
and the values are unique patient IDs for that ward.
}
